rl-tutorial/agentic_gen/requirements.txt
openai>=1.0.0
pandas
numpy
pydantic>=2.0.0
networkx>=3.0.0
python-dotenv>=1.0.0